eva_ics/hassio/requirements.txt
pydantic-settings
jinja2
pyaml
pydantic
requests
pydevd-pycharm
#pip install git+https://github.com/bashtage/sphinx-material.git
